フィルター処理の修正
